Changed test to measure fidelity rather than measurement. Moved to test_options.py
